Teamed item spawn / item knockback bugfix
Fixed a bug in which an item that shared a spawn with another item(s) could cause the next item in the team chain to spawn early, instead of having the knocked item return to its original location. Such items would either be stuck in their new positions post-knockback or even be removed from the map entirely after pickup.