Assume capture mode from client side item grab calls
The Storage Capsule is designed to be unable to store flags. This functionality has always worked as intended but had an issue where attempting to store a flag made it look like the player picked it up client-side, without actually doing so server-side. A bit of a patch-up rather than a real fix that seems to work okay for now - may find a better solution later if I can communicate storage capsule state to the client.